clean up rotTool and comment
