+++ neoexchange/core/plots.py
@@ -57,6 +57,9 @@ from photometrics.catalog_subs import sanitize_object_name
 from photometrics.SA_scatter import readSources, plotScatter, plotFormat
 from photometrics.spectraplot import spectrum_plot, read_mean_tax
 
+# JS file containing call back functions
+js_file = os.path.abspath(os.path.join('core', 'static', 'core', 'js', 'bokeh_custom_javascript.js'))
+
 logger = logging.getLogger(__name__)
 
 
@@ -598,133 +601,36 @@ def get_name(meta_dat):
     return out_string, name, number
 
 
-TS_CODE = """
-import {GestureTool, GestureToolView} from "models/tools/gestures/gesture_tool"
-import {ColumnDataSource} from "models/sources/column_data_source"
-import {PanEvent} from "core/ui_events"
-import * as p from "core/properties"
-
-function rotate(coords:any, omega:number, phi:number) {
-    const theta = 0
-    const ct = Math.cos(theta)
-    const st = Math.sin(theta)
-    const cp = Math.cos(phi)
-    const sp = Math.sin(phi)
-    const co = Math.cos(omega)
-    const so = Math.sin(omega)
-    const keys = Object.keys(coords.data) as string[]
-    const unique_keys = [] as string[]
-    var key
-    for (key of keys) {
-        const first = key.split("_")
-        if (first.length > 1 && unique_keys.indexOf(first[0]) == -1) {
-            unique_keys.push(first[0])
-            var xx = coords.get_array(first[0].concat("_x")) as number[]
-            var yy = coords.get_array(first[0].concat("_y")) as number[]
-            var zz = coords.get_array(first[0].concat("_z")) as number[]
-            for (var i = 0; i < xx.length; i++){
-                //xx[i] = xx[i] * Math.cos(theta) + yy[i] * Math.sin(theta)
-                //yy[i] = yy[i] * Math.cos(theta) - xx[i] * Math.sin(theta)
-                //yy[i] = yy[i] * Math.cos(phi) - zz[i] * Math.sin(phi)
-                //zz[i] = zz[i] * Math.cos(phi) + yy[i] * Math.sin(phi)
-                const xx_out = (xx[i] * ct * co) + (yy[i] * (ct * so * sp - st * cp)) + (zz[i] * (ct * so * cp + st * sp))
-                const yy_out = (xx[i] * st * co) + (yy[i] * (st * so * sp + ct * cp)) + (zz[i] * (st * so * cp - ct * sp))
-                const zz_out = (yy[i] * (co * sp)) + (zz[i] * (co * cp)) - (xx[i] * so)
-                xx[i] = xx_out
-                yy[i] = yy_out
-                zz[i] = zz_out
-            }
-        }
-    }
-}
-
-export class RotToolView extends GestureToolView {
-  model: RotTool
-
-  //this is executed when the pan/drag event starts
-  _pan_start(_ev: PanEvent): void {
-    this.model.source.data = {x: [], y: []}
-  }
-
-  //this is executed on subsequent mouse/touch moves
-  _pan(ev: PanEvent): void {
-    const {frame} = this.plot_view
-
-    const {sx, sy} = ev
-
-    if (!frame.bbox.contains(sx, sy))
-      return
-
-    const {source, obs, orbs} = this.model
-
-    const x_list = source.get_array("x") as number[]
-    const y_list = source.get_array("y") as number[]
-
-    var theta = 0
-    var phi = 0
-    const scale = .01
-
-    x_list.push(sx)
-    y_list.push(sy)
-
-    if (x_list.length > 2) {
-        x_list.shift()
-        y_list.shift()
-        var theta = (x_list[1] - x_list[0]) * scale
-        var phi = (y_list[1] - y_list[0]) * scale
-        }
-
-    rotate(obs, theta, phi)
-    rotate(orbs, theta, phi)
-
-    obs.change.emit()
-    orbs.change.emit()
-    source.change.emit()
-  }
-
-  // this is executed then the pan/drag ends
-  _pan_end(_ev: PanEvent): void {}
-}
-
-export namespace RotTool {
-  export type Attrs = p.AttrsOf<Props>
-
-  export type Props = GestureTool.Props & {
-    source: p.Property<ColumnDataSource>,
-    obs: p.Property<ColumnDataSource>
-    orbs: p.Property<ColumnDataSource>
-  }
-}
-
-export interface RotTool extends RotTool.Attrs {}
-
-export class RotTool extends GestureTool {
-  properties: RotTool.Props
-  __view_type__: RotToolView
-
-  constructor(attrs?: Partial<RotTool.Attrs>) {
-    super(attrs)
-  }
-
-  tool_name = "Rotate Tool"
-  icon = "bk-tool-icon-wheel-pan"
-  event_type = "pan" as "pan"
-  default_order = 12
-
-  static init_RotTool(): void {
-    this.prototype.default_view = RotToolView
-
-    this.define<RotTool.Props>({
-      source: [ p.Instance ],
-      obs: [ p.Instance ],
-      orbs: [ p.Instance ],
-    })
-  }
-}
-"""
+def get_js_as_text(file, funct):
+    """Pull out given function from js file and convert to text for Bokeh
+    Inputs:
+        file: Absolute path to js file containing desired function
+        funct: name of function to be imported into Bokeh Callback
+    """
+    with open(file, "r") as js:
+        lines = js.readlines()
+    js_text = ''
+    print_js = False
+    for line in lines:
+        if print_js and (line == '}\n' or line == '}'):
+            break
+        if print_js:
+            js_text += line
+        if "function {}(".format(funct) in line:
+            print_js = True
+    return js_text
 
 
 class RotTool(Tool):
+    """Create Bokeh Tool for tracking 3D rotation with mouse movement.
+        Add to plot with RotTool(source, obs, orbs)
+            source:..DataSource with data:(x=[] and y=[]). Used to track change in mouse position from move to move.
+            obs:.....1st data set to be looped through. Must contain equal length datasets with names formatted  as
+                     "<name>_[x,y,z]". Any columns without this formatting will be skipped, while x,y,z datasets will be
+                     transformed.
+            orbs:....2nd independent DataSource of equal length data sets using same format as obs.
+    """
+    TS_CODE = get_js_as_text(js_file, "rotation_tool")
     __implementation__ = TypeScript(TS_CODE)
     source = Instance(ColumnDataSource)
     obs = Instance(ColumnDataSource)
@@ -751,9 +657,6 @@ def lc_plot(lc_list, meta_list, period=1, pscan_dict={}, body=None, jpl_ephem=No
             ['datetime_jd'] --- Julian dates
             ['V'] --- predicted V magnitude for given Julian dates
     """
-
-    # JS file containing call back functions
-    js_file = os.path.abspath(os.path.join('core', 'static', 'core', 'js', 'bokeh_custom_javascript.js'))
 
     # Pull general info from metadata
     if body is None:
@@ -828,7 +731,6 @@ def lc_plot(lc_list, meta_list, period=1, pscan_dict={}, body=None, jpl_ephem=No
     ast_orbit = plot_orbit.line(x="asteroid_x", y="asteroid_y", source=full_orbit_source, color="gray")
     for planet in get_planetary_elements():
         plot_orbit.line(x=f"{planet}_x", y=f"{planet}_y", source=full_orbit_source, color="gray", alpha=.5)
-
     cursor_change_source = ColumnDataSource(data=dict(x=[], y=[]))
     plot_orbit.add_tools(RotTool(source=cursor_change_source, orbs=full_orbit_source, obs=orbit_source))
 
@@ -1125,23 +1027,3 @@ def translate_from_alcdef_filter(filt):
         elif filt == 'c':
             filt = 'Clear'
     return filt
-
-
-def get_js_as_text(file, funct):
-    """Pull out given function from js file and convert to text for Bokeh
-    Inputs:
-        file: Absolute path to js file containing desired function
-        funct: name of function to be imported into Bokeh Callback
-    """
-    with open(file, "r") as js:
-        lines = js.readlines()
-    js_text = ''
-    print_js = False
-    for line in lines:
-        if print_js and (line == '}\n' or line == '}'):
-            break
-        if print_js:
-            js_text += line
-        if "function {}(".format(funct) in line:
-            print_js = True
-    return js_text
